Project

+++ 0x1A-hash_tables/2-key_index.c
@@ -6,7 +6,7 @@
  * @size: array size of hash table.
  * Return: Index Key
  */
-unsigned long int key_index(const unsigned char *key, unsigned int size)
+unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
 	return (hash_djb2(key) % size);
 }
